Include DESeq analyses in snakemake pipeline
New rule deseq calls the deseq.R script, which collects data and creates DESeq objects for the 2D vs 3D and 2D vs LGG treatments.
Note that the objects are in raw DESeq format. We need to call the results function to get p-values and effect sizes.
I will implement this latter step directly in the markdown report(s).