Integrate GO enrichment in snakemake workflow
This patch converts GO enrichment rule to wildcard in order to allow us to process multiple contrasts in parallel.
I also add a new gobal variable in the snakefile, to determine the contrasts.
I use this variable in the rule all to require the output from both GO and DESeq, since they need to be consistent.
However, DESeq still uses individually defined outputs. A future patch could convert DESeq to work in parallel too, so that future experiments can run contrasts in parallel, as specified in the main snakemake file.